Following Changes are Included in This Check in 1)Add User Identity Tables and Customize Identity User Propertes 2)Add Provision For Register,Login and Assign Role 3)Add Authentication and Store Token in Cookies.

+++ Ecomm_Service.CouponAPI/Models/Coupon.cs
@@ -10,7 +10,7 @@ namespace Ecomm_Service.CouponAPI.Models
         public string CouponCode { get; set; }
         [Required]
         public double DiscountAmount { get; set; }
-        public int MinAmount { get; set; }
+        public double MinAmount { get; set; }
 
     }
 }
